+++ src/sll/api/_generated.c
@@ -790,6 +790,18 @@ __SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_object_t* sll_api_sys_get_platform_raw(con
 
 
 
+__SLL_API_TYPE_sll_api_sys_get_version sll_api_sys_get_version(__SLL_API_ARGS_sll_api_sys_get_version);
+__SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_object_t* sll_api_sys_get_version_raw(const sll_object_t*const* al,sll_arg_count_t all){
+	sll_array_t out;
+	sll_api_sys_get_version(&out);
+	sll_object_t* out_o=SLL_CREATE();
+	out_o->t=SLL_OBJECT_TYPE_ARRAY;
+	out_o->dt.a=out;
+	return out_o;
+}
+
+
+
 __SLL_API_TYPE_sll_api_sys_load_library sll_api_sys_load_library(__SLL_API_ARGS_sll_api_sys_load_library);
 __SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_object_t* sll_api_sys_load_library_raw(const sll_object_t*const* al,sll_arg_count_t all){
 	if (all<1){
@@ -913,6 +925,7 @@ static const internal_function_t _ifunc_data_ptr[]={
 	{"sll:sys_arg_get_count",sll_api_sys_arg_get_count_raw,0},
 	{"sll:sys_get_executable",sll_api_sys_get_executable_raw,0},
 	{"sll:sys_get_platform",sll_api_sys_get_platform_raw,0},
+	{"sll:sys_get_version",sll_api_sys_get_version_raw,0},
 	{"sll:sys_load_library",sll_api_sys_load_library_raw,0},
 	{"sll:time_current",sll_api_time_current_raw,0},
 	{"sll:time_current_nanos",sll_api_time_current_nanos_raw,0},
@@ -924,5 +937,5 @@ static const internal_function_t _ifunc_data_ptr[]={
 
 
 
-const sll_function_index_t _ifunc_size=48;
+const sll_function_index_t _ifunc_size=49;
 const internal_function_t* _ifunc_data=(const internal_function_t*)(&_ifunc_data_ptr);